Enable race detection in integration tests
Move tests specific to race detection to their own source file under the
race build tag.
Removed error checking for concurrent test and updated comment.
